emotion_web/emotion_web/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
deepface==0.0.79
numpy==1.24.3
requests==2.31.0
tensorflow==2.13.0
Pillow>=9.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
